requirements.txt
instagrapi==1.23.2
python-dotenv==1.0.0
requests==2.31.0